Make a client observer fixtures.
